Clean Architecture Implementation for every screen
