+++ app/src/main/java/com/basaraksanli/photoAlbum/core/presentation/components/ErrorComponent.kt
@@ -1,0 +1,18 @@
+package com.basaraksanli.photoAlbum.core.presentation.components
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+fun ErrorComponent() {
+    Box(
+        contentAlignment = Alignment.Center,
+        modifier = Modifier.fillMaxSize()
+    ) {
+        Text(text = "En error has occurred")
+    }
+}
