+++ app/src/main/java/com/basaraksanli/photoAlbum/core/presentation/components/FullScreenProgressBar.kt
@@ -1,0 +1,18 @@
+package com.basaraksanli.photoAlbum.core.presentation.components
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.CircularProgressIndicator
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+fun FullScreenProgressBar() {
+    Box(
+        contentAlignment = Alignment.Center,
+        modifier = Modifier.fillMaxSize()
+    ) {
+        CircularProgressIndicator()
+    }
+}
