+++ app/src/main/java/com/basaraksanli/photoAlbum/feature_album/presentation/albumlistscreen/AlbumListEvent.kt
@@ -1,6 +1,8 @@
 package com.basaraksanli.photoAlbum.feature_album.presentation.albumlistscreen
 
+import com.basaraksanli.photoAlbum.feature_album.domain.model.UserListItem
+
 sealed class AlbumListEvent {
     object LoadUsers : AlbumListEvent()
-    object LoadAlbums: AlbumListEvent()
+    class LoadAlbums(val userList: List<UserListItem>): AlbumListEvent()
 }
