+++ app/src/main/java/com/basaraksanli/photoAlbum/feature_album/presentation/albumlistscreen/AlbumListState.kt
@@ -1,14 +1,16 @@
 package com.basaraksanli.photoAlbum.feature_album.presentation.albumlistscreen
 
+
 import com.basaraksanli.photoAlbum.feature_album.domain.model.AlbumListItem
 import com.basaraksanli.photoAlbum.feature_album.domain.model.UserListItem
 
-data class AlbumListState(
-    val isUsersLoading : Boolean = true,
-    val isAlbumsLoading : Boolean = true,
-    val userList : List<UserListItem> = listOf(),
-    val albumList : Map<Int, List<AlbumListItem>> = mapOf(),
-    val thumbnailList : List<List<Int>> = mutableListOf(),
-    val loadError : String = "",
 
-)
+sealed class AlbumListState
+{
+    object AlbumListScreenLoading : AlbumListState()
+    class AlbumListScreenLoaded(val userList: List<UserListItem>, val albumList: Map<Int, List<AlbumListItem>>, val thumbnailList: List<List<Int>>) : AlbumListState()
+    class AlbumListScreenUsersLoaded(val userList: List<UserListItem>): AlbumListState()
+    class AlbumListScreenNetworkError(val loadError: String) : AlbumListState()
+}
+
+
