+++ app/src/main/java/com/basaraksanli/photoAlbum/feature_album/presentation/photolistscreen/PhotoListScreenEvent.kt
@@ -1,0 +1,5 @@
+package com.basaraksanli.photoAlbum.feature_album.presentation.photolistscreen
+
+sealed class PhotoListScreenEvent {
+    object LoadPhotos : PhotoListScreenEvent()
+}
