+++ app/src/main/java/com/basaraksanli/photoAlbum/feature_album/presentation/photolistscreen/PhotoListViewModel.kt
@@ -1,13 +1,16 @@
 package com.basaraksanli.photoAlbum.feature_album.presentation.photolistscreen
 
+import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
-import com.basaraksanli.photoAlbum.feature_album.domain.model.PhotoList
+import androidx.lifecycle.viewModelScope
 import com.basaraksanli.photoAlbum.feature_album.domain.model.PhotoListItem
 import com.basaraksanli.photoAlbum.feature_album.domain.use_case.AlbumUseCases
 import com.basaraksanli.photoAlbum.feature_album.util.ApiResult
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.launch
+import timber.log.Timber
 import javax.inject.Inject
 
 @HiltViewModel
@@ -15,14 +18,50 @@ class PhotoListViewModel @Inject constructor(
     private val useCases: AlbumUseCases,
     private val stateHandle: SavedStateHandle
 ) : ViewModel() {
-    var photoList = mutableStateOf<List<PhotoListItem>>(listOf())
+
+    private val _state = mutableStateOf<PhotoListScreenState>(PhotoListScreenState.PhotoListScreenLoading)
+    var state: State<PhotoListScreenState> = _state
+
     val albumTitle = stateHandle.get<String>("albumTitle")
+    private val albumId = stateHandle.get<Int>("albumId")
+    val userName = stateHandle.get<String>("userName")
 
-    suspend fun loadPhotoList(albumId: Int): ApiResult<PhotoList> {
-        return useCases.getPhotoList(albumId = albumId)
+    init {
+        onEvent(PhotoListScreenEvent.LoadPhotos)
     }
-
-    fun setPhotoList(newPhotoList: List<PhotoListItem>){
-        photoList.value = newPhotoList
+    fun onEvent(event: PhotoListScreenEvent) {
+        when (event) {
+            is PhotoListScreenEvent.LoadPhotos -> {
+                loadPhotoList(albumId!!)
+            }
+        }
+    }
+    private fun loadPhotoList(albumId: Int) {
+        viewModelScope.launch {
+            when (val result = useCases.getPhotoList(albumId = albumId)) {
+                is ApiResult.Success ->{
+                    val photoEntries = result.data!!.mapIndexed { _, entry ->
+                        PhotoListItem(
+                            albumId = entry.albumId,
+                            id = entry.id,
+                            thumbnailUrl = entry.thumbnailUrl,
+                            title = entry.title,
+                            url = entry.url,
+                        )
+                    }
+                    _state.value = PhotoListScreenState.PhotoListScreenLoaded(photoList = photoEntries)
+                }
+                is ApiResult.Error -> {
+                    if (result.message != null) {
+                        _state.value = PhotoListScreenState.PhotoListNetworkError(result.message)
+                        Timber.e(Throwable(result.message))
+                    } else {
+                        _state.value =
+                            PhotoListScreenState.PhotoListNetworkError("Unknown Error has been occurred.")
+                        Timber.e(Throwable("Unknown Error has been occurred."))
+                    }
+                }
+            }
+        }
     }
 }
