+++ app/src/main/java/com/basaraksanli/photoAlbum/feature_album/presentation/photolistscreen/components/AlbumInfoCard.kt
@@ -1,0 +1,55 @@
+package com.basaraksanli.photoAlbum.feature_album.presentation.photolistscreen.components
+
+import androidx.compose.foundation.layout.*
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun AlbumInfoCard(userName: String, albumTitle: String) {
+    Column(
+        modifier = Modifier.fillMaxHeight(),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Text(
+            text = "Album Info",
+            style = TextStyle(fontWeight = FontWeight.Bold),
+            fontSize = 24.sp,
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(vertical = 5.dp),
+            textAlign = TextAlign.Center
+        )
+        Spacer(modifier = Modifier.height(20.dp))
+        Row {
+            Column(
+                Modifier
+                    .width(110.dp)
+                    .padding(start = 20.dp),
+                verticalArrangement = Arrangement.Center
+            ) {
+                Text("User Name:", maxLines = 1, fontWeight = FontWeight.Bold)
+                Spacer(modifier = Modifier.height(10.dp))
+                Text("Album Title:", maxLines = 1, fontWeight = FontWeight.Bold)
+            }
+            Column(
+                Modifier
+                    .fillMaxSize()
+                    .padding(horizontal = 20.dp),
+            ) {
+                Text(text = userName, maxLines = 1, overflow = TextOverflow.Ellipsis)
+                Spacer(modifier = Modifier.height(10.dp))
+                Text(text = albumTitle, overflow = TextOverflow.Ellipsis)
+            }
+        }
+    }
+
+}
